Uskladjeno sa DB ver 1

+++ app/services/pricing.py
@@ -9,5 +9,9 @@ def calculate_fee(ticket, db) -> Decimal:
     minutes = max(1, int(delta.total_seconds() / 60))
     hours = max(1, math.ceil(minutes / 60))
 
-    rate = float(ticket.vehicle.vehicle_type.rate)
+    rate = Decimal(ticket.vehicle.vehicle_type.rate)
     return hours * rate
+
+
+# NOTE: This function is never called; fee is set by DB trigger.
+# It’s kept for reference, or for future use in an API-based fee calculation.
